feat: add BOM table component with metadata and cell descriptors
