+++ lib/components/BomTable/index.ts
@@ -1,0 +1,1 @@
+export * from "./BomTable"
